+++ src/Interfaces/RouteInterface.php
@@ -32,7 +32,7 @@ interface RouteInterface
     /**
      * Set callable route destination
      * 
-     * @param string $callable Function to execute for route
+     * @param Closure $callable Function to execute for route
      * @return self
      */
     public function call(Closure $callable): self;
@@ -72,44 +72,52 @@ interface RouteInterface
      * @return self 
      */
     public function params(array $params): self;
-
+    
     /**
      * Set HTTP GET routes
      * 
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
-     * @return static
+     * @return RouteInterface
      */
-    public static function get(string $path): static;
+    public static function get(string $path): RouteInterface;
 
     /**
      * Set HTTP POST routes
      * 
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
-     * @return static
+     * @return RouteInterface
      */
-    public static function post(string $path): static;
+    public static function post(string $path): RouteInterface;
 
     /**
      * Set HTTP PUT routes
      * 
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
-     * @return static
+     * @return RouteInterface
      */
-    public static function put(string $path): static;
+    public static function put(string $path): RouteInterface;
 
     /**
      * Set HTTP PATCH routes
      * 
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
-     * @return static
+     * @return RouteInterface
      */
-    public static function patch(string $path): static;
+    public static function patch(string $path): RouteInterface;
 
     /**
      * Set HTTP DELETE routes
      * 
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
-     * @return static
+     * @return RouteInterface
      */
-    public static function delete(string $path): static;
+    public static function delete(string $path): RouteInterface;
+
+    /**
+     * Set HTTP HEAD routes
+     * 
+     * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
+     * @return RouteInterface
+     */
+    public static function head(string $path): RouteInterface;
 }